+++ src/main/java/domain/score/ScorePolicy.java
@@ -4,5 +4,5 @@ import domain.Side;
 import domain.board.Board;
 
 public interface ScorePolicy {
-    int calculate(Board board, Side side);
+    double calculate(Board board, Side side);
 }
