+++ src/main/java/view/OutputView.java
@@ -65,12 +65,12 @@ public class OutputView {
                 .toList()));
     }
 
-    public void printScore(int choScore, int hanScore) {
-        System.out.printf("현재 점수 - 초: %d, 한: %d%n", choScore, hanScore);
+    public void printScore(double choScore, double hanScore) {
+        System.out.printf("현재 점수 - 초: %.1f, 한: %.1f%n", choScore, hanScore);
     }
 
-    public void printFinalScore(int choScore, int hanScore) {
-        System.out.printf("최종 점수 - 초: %d, 한: %d%n", choScore, hanScore);
+    public void printFinalScore(double choScore, double hanScore) {
+        System.out.printf("최종 점수 - 초: %.1f, 한: %.1f%n", choScore, hanScore);
     }
 
     public void printWinner(String winner) {
